+++ CHANGELOG.md
@@ -6,7 +6,11 @@
 
 - [language] feature: now the `multi-language` feature supports `MC 1.20` and `MC 1.20.1`.
 - [nametag] Enhance this module
+  - docs: document how this module works.
   - feature: improve the responsiveness of the nametag hiding when press `SHIFT` key.
   - feature: use a better implementation.
     - Before: The nametag entity has to ride the owner, to sync the position update.
-    - Now: Use the virtual riding, and simulate the riding at client-side, making it better compatibility with vanilla behaviours and other mods.
+    - **Now: Use the virtual riding, and simulate the riding at client-side, making it better compatibility with vanilla behaviours and other mods.**
+  - feature: new `/nametag toggle` command, to toggle between `fuji nametag` and `vanilla nametag`.
+    - Example: `/nametag toggle @s`
+    - Example: `/nametag toggle others @a false`